Compilation fix for gcc 6.2.0

+++ radio/src/gvars.cpp
@@ -74,7 +74,7 @@ int16_t getGVarValue(int8_t gv, int8_t fm)
 int32_t getGVarValuePrec1(int8_t gv, int8_t fm)
 {
   int8_t mul;
-  uint8_t prec = g_model.gvars[abs(gv)].prec;
+  uint8_t prec = g_model.gvars[abs((int)gv)].prec;    // explicit cast to `int` needed, othervise gv is promoted to double!
   if (prec == 0)
     mul = 10;
   else
